[#68] chore : type 정리
