+++ src/types/profile.types.ts
@@ -1,0 +1,58 @@
+export interface USERPROFILE {
+  email: string;
+  username: string;
+  profileImg: string;
+  phone: string;
+  trackName: string;
+  trackRole: string;
+  period: string;
+}
+
+export interface UserInfoProps {
+  username: string;
+  trackName: string;
+  period: string;
+  trackRole: string;
+  profileImg: string;
+  handleChangeImage: () => void;
+  handleChangeRole: () => void;
+}
+
+export interface AccountInfoProps {
+  email: string;
+  phone: string;
+  handleChangePassword: () => void;
+  handleChangePhoneNumber: () => void;
+}
+
+//모달
+export interface ChangeModalProps {
+  onClose: () => void;
+}
+
+export interface RoleChangeModalProps {
+  trackRole: string;
+  onClose: () => void;
+}
+
+export interface RandomProfileModalProps {
+  onClose: () => void;
+  profileImg: string;
+}
+
+//api
+export interface PassWordChange {
+  originPassword: string;
+  newPassword: string;
+}
+
+export interface RoleChange {
+  trackName: string;
+  period: string;
+  trackRole: string;
+}
+
+export interface ProfileChange {
+  profileImg: string;
+  nickname: string;
+}
